refactor(memory): keep admission metadata a leaf import
Storage now imports the vocabulary from a zero-dependency module so
lightweight Memory-runtime contract checks do not load CaptureAdmission
or aiohttp.
